pyproject.toml
[project]
name = "agent-gate"
version = "0.1.0"
description = "An execution gateway for AI agents on untrusted devices"
requires-python = ">=3.12"
license = "MIT"
dependencies = [
    "websockets>=14.0,<17.0",
    "python-telegram-bot[callback-data]>=21.0,<22.0",
    "aiosqlite>=0.20.0,<1.0",
    "aiohttp>=3.10.0,<4.0",
    "pyyaml>=6.0,<7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0",
    "ruff>=0.9.0",
    "pre-commit>=4.0",
]

[project.scripts]
agent-gate = "agent_gate.__main__:main"

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort
    "UP",    # pyupgrade
    "B",     # flake8-bugbear
    "SIM",   # flake8-simplify
    "RUF",   # ruff-specific
]

[tool.ruff.lint.isort]
known-first-party = ["agent_gate"]
